扇形图

+++ app/src/main/java/com/example/admin/mycircleview/PieChartData.kt
@@ -9,10 +9,10 @@ class PieChartData {
 
     var color: Int? = null
     var name: String? = null
-    var num: Int? = null
+    var num: Float? = null
     var unit:String?=null
 
-    constructor(color: Int?, name: String?, num: Int?,unit:String?) {
+    constructor(color: Int?, name: String?, num: Float?,unit:String?) {
         this.color = color
         this.name = name
         this.num = num
